requirements.txt
accelerate==1.6.0
altair==5.5.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.1
confection==0.1.4
contexttimer==0.3.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
debugpy==1.8.0
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.16.0
distlib==0.3.8
editdistance==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.4
fastjsonschema==2.19.1
filelock==3.13.1
fire==0.6.0
fonttools==4.47.0
fqdn==1.5.1
fsspec==2023.12.2
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.31.2
identify==2.5.33
idna==3.6
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==6.11.0
importlib-resources==6.1.1
inspectus==0.2.0
iopath==0.1.10
ipykernel==6.28.0
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.0
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaggle==1.5.16
kiwisolver==1.4.5
labml==0.5.3
langcodes==3.3.0
lazy_loader==0.3
Levenshtein==0.23.0
lmdb==1.5.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
murmurhash==1.0.10
narwhals==1.28.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai-clip==1.0.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
paddleocr==2.8.1
paddlepaddle-gpu==2.6.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
peft==0.7.0
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.1.0
plotly==5.18.0
portalocker==2.8.2
pre-commit==3.6.0
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
PuLP==2.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyclipper==1.3.0.post5
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.2
python-json-logger==2.0.7
python-Levenshtein==0.23.0
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.6.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.16.2
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.2
sentencepiece==0.1.99
shapely==2.0.5
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.29.0
style==1.1.0
sympy==1.14.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.38.0
triton==3.3.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
update==0.0.1
uri-template==1.3.0
urllib3==2.1.0
validators==0.22.0
virtualenv==20.25.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.12
weasel==0.3.4
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
zipp==3.17.0
